UpperConfidenceBound(UCB)
Here we used Upper Confidence Bound and random selection to predict the best ad out of 10 we call it multi bandit problem. Use marketing department as the best idea to predict what is the best version of the Ad that will most likely clicked by people using this UCB program and we used random selection method to compare two algorithms. Of course We used R to program in this case.