UCBVSRandomSelectionPlots
We will use this plot to visualize which ad version is mostly selected or clicked online. It is very clear, In this case it is ad version 5 is mostly clicked. and random selection is not a good algorithm for this particular problem as it randomly selected the ad version for each user with poor performance.